requirements.txt
absl-py==2.1.0
addict==2.4.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
easydict==1.11
einops==0.7.0
exceptiongroup==1.2.2
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
google-auth==2.32.0
google-auth-oauthlib==1.0.0
grpcio==1.65.1
huggingface-hub==0.24.2
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
iniconfig==2.0.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
mmcv-full==1.7.1
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
protobuf==5.27.2
psutil==5.8.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pytest==7.4.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==0.9.8
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tornado==6.4.1
tqdm==4.66.4
ttach==0.0.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visdom==0.2.4
websocket-client==1.8.0
Werkzeug==3.0.3
yacs==0.1.8
yapf==0.40.2
zipp==3.19.2
